src/third_party/blink/web_tests/http/tests/security/xssAuditor/script-tag-inside-svg-tag5-expected.txt

CONSOLE ERROR: line 4: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/echo-intertag.pl?clutter=blah&q=%3Csvg%3E%3Cscript%3Ealert(0)//&q2=%3C/script%3E%3C/svg%3E' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
 Ensures script blocks in a context where CDATA is allowed respect JS comments.
